+++ src/test/java/someMathTest/RationalNumberTest.java
@@ -17,51 +17,51 @@ public class RationalNumberTest
 	{
 
 		RationalNumber rn = 
-				new RationalNumber(false, new NaturalNumber(0), new NaturalNumber(88), new NaturalNumber(88));
+				new RationalNumber(false, new NaturalNumber((long)(long)0), new NaturalNumber((long)(long)88), new NaturalNumber((long)(long)88));
 
 		RationalNumber rn2 = 
-				new RationalNumber(false, new NaturalNumber(0), new NaturalNumber(2), new NaturalNumber(2));
+				new RationalNumber(false, new NaturalNumber((long)(long)0), new NaturalNumber((long)(long)2), new NaturalNumber((long)(long)2));
 	
 		System.out.println(rn + " vs " + rn2);
 		
 		assert(rn.equals(rn2));
 		
 		RationalNumber rn3 = 
-				new RationalNumber(true, new NaturalNumber(9), new NaturalNumber(88), new NaturalNumber(88));
+				new RationalNumber(true, new NaturalNumber((long)9), new NaturalNumber((long)88), new NaturalNumber((long)88));
 
 		RationalNumber rn4 = 
-				new RationalNumber(true, new NaturalNumber(5), new NaturalNumber(10), new NaturalNumber(2));
+				new RationalNumber(true, new NaturalNumber((long)5), new NaturalNumber((long)10), new NaturalNumber((long)2));
 	
 		System.out.println(rn3 + " vs " + rn4);
 		
 		assert(rn3.equals(rn4));
 
 		RationalNumber rn5 = 
-				new RationalNumber(false, new NaturalNumber(9), new NaturalNumber(11), new NaturalNumber(88));
+				new RationalNumber(false, new NaturalNumber((long)9), new NaturalNumber((long)11), new NaturalNumber((long)88));
 		RationalNumber rn6 = 
-				new RationalNumber(false, new NaturalNumber(8), new NaturalNumber(99), new NaturalNumber(88));
+				new RationalNumber(false, new NaturalNumber((long)8), new NaturalNumber((long)99), new NaturalNumber((long)88));
 
 		System.out.println(rn5 + " vs " + rn6);
 
 		assert(rn5.equals(rn6));
 		
 		RationalNumber rn7 = 
-				new RationalNumber(false, new NaturalNumber(9), new NaturalNumber(11), new NaturalNumber(77));
+				new RationalNumber(false, new NaturalNumber((long)9), new NaturalNumber((long)11), new NaturalNumber((long)77));
 		RationalNumber rn8 = 
-				new RationalNumber(false, new NaturalNumber(8), new NaturalNumber(88), new NaturalNumber(77));
+				new RationalNumber(false, new NaturalNumber((long)8), new NaturalNumber((long)88), new NaturalNumber((long)77));
 
 		System.out.println(rn7 + " vs " + rn8);
 
 		assert(rn7.equals(rn8));
 
 		RationalNumber rn9 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn10 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 
 		RationalNumber sum = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(2), new NaturalNumber(7*9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)2), new NaturalNumber((long)7*9));
 
 
 		System.out.println(rn9 + " + " + rn10 + " = " + sum);
@@ -69,60 +69,60 @@ public class RationalNumberTest
 		assert(sum.equals(rn9.add(rn10)));
 		
 		RationalNumber rn11 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn12 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		sum = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(2), new NaturalNumber(7*9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)2), new NaturalNumber((long)7*9));
 
 		System.out.println(rn11 + " + " + rn12 + " = " + sum);
 
 		assert(sum.equals(rn11.add(rn12)));
 
 		RationalNumber rn13 = 
-				new RationalNumber(true, new NaturalNumber(2), new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(true, new NaturalNumber((long)2), new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn14 = 
-				new RationalNumber(false, new NaturalNumber(3), new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(false, new NaturalNumber((long)3), new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		sum = 
-				new RationalNumber(false, NaturalNumber.one, new NaturalNumber(2), new NaturalNumber(7*9));
+				new RationalNumber(false, NaturalNumber.one, new NaturalNumber((long)2), new NaturalNumber((long)7*9));
 
 		System.out.println(rn13 + " + " + rn14 + " = " + sum);
 
 		assert(sum.equals(rn13.add(rn14)));
 		
 		RationalNumber rn15 = 
-				new RationalNumber(false, new NaturalNumber(2), new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(false, new NaturalNumber((long)2), new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn16 = 
-				new RationalNumber(true, new NaturalNumber(3), new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(true, new NaturalNumber((long)3), new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		sum = 
-				new RationalNumber(true, NaturalNumber.one, new NaturalNumber(2), new NaturalNumber(7*9));
+				new RationalNumber(true, NaturalNumber.one, new NaturalNumber((long)2), new NaturalNumber((long)7*9));
 
 		System.out.println(rn15 + " + " + rn16 + " = " + sum);
 
 		assert(sum.equals(rn15.add(rn16)));
 		
 		RationalNumber rn17 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn18 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		sum = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(9+7), new NaturalNumber(7*9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)9+7), new NaturalNumber((long)7*9));
 
 		System.out.println(rn17 + " + " + rn18 + " = " + sum);
 
 		assert(sum.equals(rn17.add(rn18)));
 
 		RationalNumber rn19 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn20 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		sum = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(9+7), new NaturalNumber(7*9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)9+7), new NaturalNumber((long)7*9));
 
 		System.out.println(rn19 + " + " + rn20 + " = " + sum);
 
@@ -148,12 +148,12 @@ public class RationalNumberTest
 		assert(rn25.equals(rn26));
 
 		RationalNumber rn27 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		RationalNumber rn28 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		RationalNumber product = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7*9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7*9));
 
 
 		System.out.println(rn27 + " * " + rn28 + " = " + product);
@@ -161,33 +161,33 @@ public class RationalNumberTest
 		assert(product.equals(rn27.multiply(rn28)));
 		
 		rn27 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		rn28 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		product = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7*9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7*9));
 
 		System.out.println(rn27 + " * " + rn28 + " = " + product);
 
 		assert(product.equals(rn27.multiply(rn28)));
 
 		rn27 = 
-				new RationalNumber(true, new NaturalNumber(2), new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(true, new NaturalNumber((long)2), new NaturalNumber((long)1), new NaturalNumber((long)9));
 		rn28 = 
-				new RationalNumber(false, new NaturalNumber(3), new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(false, new NaturalNumber((long)3), new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		product = 
-				new RationalNumber(false, new NaturalNumber(19*22), new NaturalNumber(7*9));
+				new RationalNumber(false, new NaturalNumber((long)19*22), new NaturalNumber((long)7*9));
 
 		System.out.println(rn27 + " * " + rn28 + " = " + product);
 
 		assert(product.equals(rn27.multiply(rn28)));
 		
 		rn27 = 
-				new RationalNumber(false, new NaturalNumber(2), new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(false, new NaturalNumber((long)2), new NaturalNumber((long)1), new NaturalNumber((long)9));
 		rn28 = 
-				new RationalNumber(true, new NaturalNumber(3), new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(true, new NaturalNumber((long)3), new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		//Product already there
 		System.out.println(rn27 + " * " + rn28 + " = " + product);
@@ -195,24 +195,24 @@ public class RationalNumberTest
 		assert(product.equals(rn27.multiply(rn28)));
 		
 		rn27 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		rn28 = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		product = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7*9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7*9));
 
 		System.out.println(rn27 + " * " + rn28 + " = " + product);
 
 		assert(product.equals(rn27.multiply(rn28)));
 
 		rn27 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(9));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)9));
 		rn28 = 
-				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7));
+				new RationalNumber(false, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7));
 
 		product = 
-				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber(1), new NaturalNumber(7*9));
+				new RationalNumber(true, NaturalNumber.zero, new NaturalNumber((long)1), new NaturalNumber((long)7*9));
 
 		System.out.println(rn27 + " * " + rn28 + " = " + product);
